styles and layout done
