+++ package.json
@@ -29,7 +29,9 @@
     "eslintConfig": {
         "extends": [
             "react-app",
-            "react-app/jest"
+            "react-app/jest",
+            "airbnb/base",
+            "plugin:prettier/recommended"
         ]
     },
     "browserslist": {
