+++ src/components/buttons/Buttons.module.css
@@ -1,12 +1,12 @@
 .buttons {
     display: flex;
-    width: 300px;
+    width: 400px;
     height: 30px;
     justify-content: space-around;
-    position: fixed;
-    top: 26%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    align-items: center;
+    margin-top: 20px;
+    margin-left: auto;
+    margin-right: auto;
 }
 
 button {
