+++ src/components/countSteps/CountSteps.module.css
@@ -1,7 +1,9 @@
 .numbersteps {
-    position: fixed;
+    display: block;
+    text-align: center;
     font-size: 20px;
-    top: 28%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    width: 400px;
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 20px;
 }
