+++ src/components/field/Field.module.css
@@ -1,10 +1,9 @@
 .field {
     display: flex;
     flex-wrap: wrap;
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 20px;
 }
 
 .cells {
