+++ src/components/header/Header.module.css
@@ -1,6 +1,8 @@
 .namegame {
-    position: fixed;
-    top: 20%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    display: block;
+    width: 400px;
+    text-align: center;
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 200px;
 }
